+++ Source/Cells/ViewControllerCell.swift
@@ -1,7 +1,7 @@
 // MIT license. Copyright (c) 2018 SwiftyFORM. All rights reserved.
 import UIKit
 
-public class ViewControllerFormItemCellModel {
+public class ViewControllerCellModel {
 	public let title: String
 	public let placeholder: String
 	public init(title: String, placeholder: String) {
@@ -10,11 +10,11 @@ public class ViewControllerFormItemCellModel {
 	}
 }
 
-public class ViewControllerFormItemCell: UITableViewCell, SelectRowDelegate {
-	public let model: ViewControllerFormItemCellModel
-	let innerDidSelectRow: (ViewControllerFormItemCell, ViewControllerFormItemCellModel) -> Void
+public class ViewControllerCell: UITableViewCell, SelectRowDelegate {
+	public let model: ViewControllerCellModel
+	let innerDidSelectRow: (ViewControllerCell, ViewControllerCellModel) -> Void
 
-	public init(model: ViewControllerFormItemCellModel, didSelectRow: @escaping (ViewControllerFormItemCell, ViewControllerFormItemCellModel) -> Void) {
+	public init(model: ViewControllerCellModel, didSelectRow: @escaping (ViewControllerCell, ViewControllerCellModel) -> Void) {
 		self.model = model
 		self.innerDidSelectRow = didSelectRow
 		super.init(style: .value1, reuseIdentifier: nil)
@@ -36,3 +36,9 @@ public class ViewControllerFormItemCell: UITableViewCell, SelectRowDelegate {
 		SwiftyFormLog("did invoke")
 	}
 }
+
+@available(*, unavailable, renamed: "ViewControllerCell")
+typealias ViewControllerFormItemCell = ViewControllerCell
+
+@available(*, unavailable, renamed: "ViewControllerCellModel")
+typealias ViewControllerFormItemCellModel = ViewControllerCellModel